devex-959: move IrsNotice out a level so not specific to CaseDetailEdit
